third_party/blink/web_tests/virtual/web-components-v0-disabled/shadow-dom/window-event-expected.txt

CONSOLE ERROR: line 14: Uncaught TypeError: shadowHostV0.createShadowRoot is not a function
This is a testharness.js-based test.
FAIL window-event Uncaught TypeError: shadowHostV0.createShadowRoot is not a function
Harness: the test ran to completion.
